// src/types/admin-card.ts
export type AdminCard = {
  _id?: string;
  bankName: string;
  joiningAnnualFee: string;
  apr: string;
  salaryTransferRequired: boolean;
  welcomeBonus: string;
  earnRates: string;
  keyLifestyleBenefits: string;
  pointsRedemption: string;
  documentsRequired: string;
  cardImageUrl: string;
  createdAt?: string;
  updatedAt?: string;
};